Moved constant to root and created utils for calculate winner
