+++ react/tic-tac-toe/src/constants/index.js
